+++ packages/babel-preset-expo/src/server-actions-plugin.ts
@@ -124,7 +124,7 @@ export function reactServerActionsPlugin(
     );
 
     // Create a top-level declaration for the extracted function.
-    const bindingKind = 'const';
+    const bindingKind = 'var';
     const functionDeclaration = t.exportNamedDeclaration(
       t.variableDeclaration(bindingKind, [
         t.variableDeclarator(extractedIdentifier, extractedFunctionExpr),
@@ -386,7 +386,7 @@ export function reactServerActionsPlugin(
               const id = left;
               const exportedSpecifier = t.exportSpecifier(id, t.identifier('default'));
               // Replace `export default foo = async () => {}` with `const foo = async () => {}`
-              path.replaceWith(t.variableDeclaration('const', [t.variableDeclarator(id, right)]));
+              path.replaceWith(t.variableDeclaration('var', [t.variableDeclarator(id, right)]));
               // Insert `(() => _registerServerReference(foo, "file:///unknown", "default"))();`
               path.insertAfter(t.exportNamedDeclaration(null, [exportedSpecifier]));
             } else if (
@@ -400,7 +400,7 @@ export function reactServerActionsPlugin(
               const extractedIdentifier = moduleScope.generateUidIdentifier('$$INLINE_ACTION');
 
               // @ts-expect-error: Transform `export default async () => {}` to `const $$INLINE_ACTION = async () => {}`
-              path.node.declaration = t.variableDeclaration('const', [
+              path.node.declaration = t.variableDeclaration('var', [
                 t.variableDeclarator(extractedIdentifier, path.node.declaration),
               ]);
               // Strip the `export default`